refactor(factories): move tests to dedicated test/ folders
