+++ packages/factories/src/factories/test/user.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest';
-import { MemoryAdapter } from '../memory-adapter';
-import { createUserFactory } from './user';
+import { MemoryAdapter } from '../../memory-adapter';
+import { createUserFactory } from '../user';
 
 describe('createUserFactory', () => {
   it('builds a UserDTO with required fields', () => {
